Refactor storage related classes into storage package
