+++ src/main/java/seedu/duke/Duke.java
@@ -4,6 +4,7 @@ import seedu.duke.event.EventManager;
 import seedu.duke.exception.InvalidValueException;
 import seedu.duke.parser.CommandParser;
 import seedu.duke.parser.CommandType;
+import seedu.duke.storage.StorageManager;
 
 import java.io.IOException;
 import java.util.ArrayList;
