+++ src/main/java/seedu/duke/storage/EventListEncoder.java
@@ -1,4 +1,9 @@
-package seedu.duke;
+package seedu.duke.storage;
+
+import seedu.duke.Cca;
+import seedu.duke.Class;
+import seedu.duke.Event;
+import seedu.duke.Test;
 
 import java.util.ArrayList;
