+++ src/main/java/seedu/duke/storage/StorageManager.java
@@ -1,5 +1,9 @@
-package seedu.duke;
+package seedu.duke.storage;
 
+import seedu.duke.Cca;
+import seedu.duke.Class;
+import seedu.duke.Event;
+import seedu.duke.Test;
 import seedu.duke.exception.InvalidValueException;
 
 import java.io.File;
