+++ src/test/java/seedu/duke/StorageTest.java
@@ -2,6 +2,7 @@ package seedu.duke;
 
 import org.junit.jupiter.api.Test;
 import seedu.duke.exception.InvalidValueException;
+import seedu.duke.storage.StorageManager;
 
 import java.util.ArrayList;
